Track separate network- and app-level byte counts
M-Lab is updating things on their end to have two separate bytes
sent/received counts: one for everything sent over TCP, including TLS
and WebSocket overheads, and one for bytes sent in WebSocket payloads
only. The former is labeled "network" (although it feels more like
transport-layer to me), and the later "application".

Track the "network" bytes sent/received in addition to the "application"
bytes sent and received. Use the network counts for metrics as they more
accurately reflect the actual throughput of the network.

For now, accept and send a measurement object compatible with both
schemas (with only one set of counts or with two) since the change has
been implemented in M-Lab's MSAK server code, which we use for local
testing, and deployed to their staging servers but has not yet been
deployed to production.