Handle slow-to-start throughput streams
Sometimes one or more of the throughput streams are slow to start. If
the WebSocket isn't opened before the end of the warmup period, the test
will fail, but the WebSocket doesn't seem to get closed correctly. Fix
that problem, add some logs to make detecting this issue easier, and
extend the maximum warmup time to reduce its likelihood.