Use device ID and secret to secure DB access
Implement a new auth flow that will limit each device to only accessing
or inserting data tied to its device ID. Corresponding DB changes,
including enabling row-level security and creating the registration
function, are documented in the server repository.

Before each interaction with the Supabase DB, check whether we have a
device secret stored. If not, register with the DB to get a secret.
Store the secret encrypted in the local data store.

Include the device ID and secret as HTTP headers on every request to
Supabase, allowing the DB to verify that we are the original owner of
the device ID.

The database only stores a hashed version of the device secret. There
remain two ways for an attacker to obtain the secret and thereby insert
or read data corresponding to a device ID that it hasn't registered:

1. By breaking into Android's KeyStore to get the encryption key used
   when storing the secret locally.
2. By compromising the HTTPS connection between the device and Supabase
   and extracting the header.

These attack vectors both require defeating existing, time- and
battle-tested security mechanisms.

As a result of these changes, the Supabase anon key that the app uses no
longer needs to be treated as a secret. It still won't be shared
publicly, but the risk factor from someone extracting it from an APK is
now significantly lower.

Note that anyone with the anon key could register a device ID and insert
fake data. However, they would only be able to read data tied to a
device ID that they register. Additionally, any data that they insert
must be tied to a device ID that they've registered and therefore could
be easily removed once identified.