Show measurement groups with MapBox's clustering
Show measurement groups instead of individual measurements on the map.
Use the first location from the group as the location for now -- we'll
likely change this behavior at some point in the future.

Also switch to MapBox's native clustering. Our custom clustering is
insensitive to the zoom level and slow. MapBox has clustering
capabilities of its own that seem to work much better; use those
instead.

Fixes: #26