Show measurement groups in all relevant hexes
A measurement group may have components in multiple hexagons. For
example, the latency component may have occurred in one hex, but the
download and upload components occurred in a different one. When this
happens, show the group in all relevant hexes, with a warning shown for
the components that don't belong in that hex.