Show measurements as points when zoomed out
Rendering the hex grid when zoomed out doesn't make much sense, but it's
nice to still be able to see your measurements. If the hex grid is
turned on, switch to the points view (without changing the toggle
switch) when zooming out, and auto-switch back to the hexes (again,
without changing the toggle switch) when zooming back in. If the hex
grid is turned off, keep the points view always as expected.