Calculate throughput as sum of streams' throughput
Since we calculate bytes transferred based on the MSAK measurement
messages, which are a snapshot in time, the most accurate way to
calculate throughput is to also use the elapsed time values in the
measurement messages. In order to do so, we must calculate throughput
for each stream independently and then sum the results. This approach
should be more accurate than summing the bytes transferred and then
using a single duration value.